Make the core, tools, and LLM packages strict-mypy clean
Annotate the streaming generator protocol in llm/base, the tool registry's
permission flow, loop tool-call handling, and the remaining tools so the
whole package passes mypy --strict without ignores. Behavior is unchanged;
types now match what the runtime already guaranteed.

+++ aegisx_agent/tools/coding/codebase.py
@@ -305,7 +305,7 @@ class CodebaseTool(Tool):
         return ToolResult(status=ToolStatus.SUCCESS, output=output)
 
     @staticmethod
-    def _format_size(size: int) -> str:
+    def _format_size(size: float) -> str:
         for unit in ["B", "KB", "MB"]:
             if size < 1024:
                 return f"{size:.0f}{unit}"
